transfer-learning/format-convertor/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
safetensors>=0.4.2
mlx>=0.5.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
llama-cpp-python>=0.2.56  # For GGUF support
